refactor: apply post like review feedback
